Implement hard link support
This change adds support for hard links, and tests to go along with
that.

There is a bit of ugliness in this change around the use of the name
field on FakeFile instances. Support for hard links means that the same
file can have different names under different directories, so it
fundamentally doesn't make sense for FakeFile instances to have a name
associated with them.

However, I didn't want to have to perform a mass refactoring as a
prerequisite to making this change, so it instead just sets the filename
as needed to work with the existing implementation of AddObject. This
has the unavoidable side effect of changing the name associated with the
original file as well, since they share the same FakeFile instance.

It looks like this is probably safe to do now, but to properly ensure
that there aren't any problems related to this, the whole codebase
should be refactored to remove the name field from FakeFile, and replace
it with a suitable alternative. It may require a wrapper class that
helps callers keep track of a file's name while preventing misuse of
that information by direct users of FakeFile.