Add FakeFileOpen.__name__, which pytest uses in messages
Having said that, pytest's messages are INTERNAL ERRORs, so this is
only the beginning of a solution to #43.  But it does make the problem
clearer.

Run pytest as instructed at the beginning of file pytest_doctest_test.py.

Earlier, I misinterpreted @megapctr's statement that this "crashes".  I was
looking for extreme drama like a segmentation fault.  Having said that,
INTERNAL ERRORs are certainly bad.