Fix shutil.move implementation
If both the source and destination argument of shutil.move are
directories the source directory is copied under the target directory.
pyfakefs implementation copied the source directory to the target
directory.

The real shutil.move call shutil.copy('/path/to/src', '/path/to/dest')
moves /path/to/src to /path/to/dest/src but the pyfakefs implementation
moved the src directory to /path/to/dest