Fix scandir DirEntry.path for relative path or fd
The documentation for os.DirEntry.path [1] says that the path attribute
of a DirEntry object is only absolute if scandir()'s path argument was
absolute, and simply the name attribute if scandir()'s path argument was
a filedescriptor. This patch makes pyfakefs do this, and adjusts the
tests.

Relative symlinks are now also covered in the scandir tests.

RealScandirFdTest.test_path_like was failing for me in Python 3.7.3 on
Debian Bullseye.

[1] https://docs.python.org/3/library/os.html#os.DirEntry.path