Avoid relative path corner-cases in scandir tests.
On Windows the initial cwd of the fake filesystem seems to be \ [3] and
the base_path of the fixture is C:\basepath [4]. I don't know what the
relative paths from \ to paths below C:\basepath are, since Windows
seems to have per-drive roots instead of a single one [2]. If it's
C:\basepath\..., it's actually an absolute path on Windows, negating the
purpose of the tests that use relative paths.

On non-Windows systems, the initial cwd of the fake filesystem seems to
be /, so that the relative paths are relative to the root. That might
hide problems where the current working directory is not taken into
account and instead paths are interpreted relative to the root, lowering
the value of the tests that use relative paths.

By changing the working directory to self.base_path in setUp, this patch
let's us avoid problems like the above.

[2] https://docs.microsoft.com/en-us/dotnet/standard/io/file-path-formats
[3] https://github.com/jmcgeheeiv/pyfakefs/blob/7e6f845a795ee6be7783f15033b623f08b2707ed/pyfakefs/fake_filesystem.py#L894
[4] https://github.com/jmcgeheeiv/pyfakefs/blob/7e6f845a795ee6be7783f15033b623f08b2707ed/pyfakefs/tests/test_utils.py#L344-L346